gts3l-common: fs: Import firmware_mnt files from stock firmware
I don't want to dump all these files like this way, but without this the
filesystem mounting doesn't work properly so that the keymaster couldn't
get the skeymast files from the /vendor/firmware_mnt/image folder for a
while. Rediculously, it got the files after few minutes. Why?

This leads boot loops and make Bluetooth keep crash since keystore is
not loaded properly because of the reason above.

I don't know why, but simply dumping the firmwares and import them will
fix this issue.

The ROM size will be increased so much, but there is no other way to fix
this, for now.

Signed-off-by: Deokgyu Yang <secugyu@gmail.com>
Change-Id: Ibe5be05233cc31ea190247dd65441170f87acba9